requirements.txt
absl-py==2.2.2
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
astunparse==1.6.3
attrs==25.3.0
azure-core==1.34.0
beautifulsoup4==4.12.3
blinker==1.9.0
bs4==0.0.2
cachetools==5.5.2
certifi==2024.7.4
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.8
cloudpickle==3.1.1
colorama==0.4.6
contourpy==1.2.1
cryptography==45.0.4
cycler==0.12.1
defusedxml==0.7.1
distro==1.9.0
Django==5.0.6
docutils==0.21.2
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
flatbuffers==25.2.10
fonttools==4.53.1
fpdf==1.7.2
fpdf2==2.8.2
gast==0.6.0
git-filter-repo==2.47.0
glm==0.4.4
google-auth==2.40.3
google-genai==1.21.1
google-pasta==0.2.0
greenlet==3.2.3
grpcio==1.71.0
guard==1.0.1
gym==0.26.2
gym-notices==0.0.8
h11==0.16.0
h5py==3.13.0
html5lib==1.1
httpcore==1.0.9
httpx==0.28.1
id==1.5.0
idna==3.7
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
Jinja2==3.1.6
jiter==0.10.0
keras==3.9.2
keyring==25.6.0
kiwisolver==1.4.5
libclang==18.1.1
listrindex==0.1
livereload==2.7.1
lxml==5.3.0
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.1.post1
mdurl==0.1.2
ml_dtypes==0.5.1
more-itertools==10.7.0
MouseInfo==0.1.3
mpmath==1.3.0
mysql-connector==2.2.9
mysql-connector-python==9.1.0
mysql-connector-python-dd==2.0.2
namex==0.0.9
nh3==0.2.21
numpy==2.0.0
openai==1.91.0
opt_einsum==3.4.0
optree==0.15.0
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.2
pdfminer.six==20250506
pdfplumber==0.11.7
pillow==10.4.0
protobuf==5.29.4
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAutoGUI==0.9.54
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pygame==2.6.1
PyGetWindow==0.0.9
Pygments==2.19.1
PyMsgBox==1.0.9
PyMySQL==1.1.1
pyparsing==3.1.2
pypdfium2==4.30.1
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytweening==1.2.0
pytz==2024.1
pywhatkit==5.4
pywin32-ctypes==0.2.3
randomuser==1.6
readme_renderer==44.0
reportlab==4.4.1
requests==2.32.3
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==14.0.0
rsa==4.9.1
selenium==4.29.0
setuptools==80.3.1
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.41
sqlparse==0.5.0
sympy==1.13.2
tenacity==8.5.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
termcolor==3.1.0
tornado==6.5.1
tqdm==4.67.1
trio==0.29.0
trio-websocket==0.12.2
twine==6.1.0
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
wheel==0.45.1
wikipedia==1.4.0
wrapt==1.17.2
wsproto==1.2.0
